refactoring tcell functions finished
tcell functions were moved around and included in gocui code directly.
`tcell_driver.go` file was limited to only specific tcell functions
which needs special transform (like Poll or SetCell).
